+++ cmd/agentbox-server/main.go
@@ -19,6 +19,7 @@ import (
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/health"
 	healthv1 "google.golang.org/grpc/health/grpc_health_v1"
+	"google.golang.org/grpc/keepalive"
 )
 
 func main() {
@@ -106,7 +107,7 @@ func run(logger *slog.Logger, configPath string) error {
 		ReadHeaderTimeout: 10 * time.Second,
 		IdleTimeout:       90 * time.Second,
 	}
-	grpcServer := grpc.NewServer()
+	grpcServer := grpc.NewServer(grpc.KeepaliveEnforcementPolicy(keepalive.EnforcementPolicy{MinTime: 5 * time.Second, PermitWithoutStream: true}))
 	controlPlane.RegisterGRPC(grpcServer)
 	healthServer := health.NewServer()
 	healthv1.RegisterHealthServer(grpcServer, healthServer)